Another example of using a class variable as default that doesn't work as expected with inheritance
This is probably why we usually had a hard-time setting the free_gas_strategy in the past.
Anyway, it also closes #2421 as now BlockchainInterface.gas_strategy always has a value; previously it was possible to pass None in the constructor (e.g. if the config file had an explicit "null" value).