Added an exists check for the user cache folder
This is necessary for the run_demo_ursula_fleet.py script in the
examples folder. It's frustrating for new users to follow documentation
and have an error like this pop up. Additionally, the error made it
seem like sage.db is a file. Because there is a product called sage.db,
this caused even more confusion. Now the correct folder is created if
it does not already exist.