Rebase fixes for rebased-chili-limes (#28)
* Fix PRETask serialization

* Use a constant for encrypted kfrag length

* Fix argument order in Signature.verify()

* Remove a too brittle check in test_bob_handles_frags

What exactly are we ensuring here? If we want to check that the cfrag is fine,
we need to use it for decryption.

* Remove an unnecessary check

The result of `reencrypt()` is verified by default.

* Re-raise Umbral decryption errors as DecryptionFailed

* Fix an argument of _filter_work_orders_and_capsules()

* Fix an assertion in test_federated_grant_and_revoke

Assertion fails because now UmbralMessageKit has __len__,
so a bool() call is equivalent to len() == 0
(which it is, since there are no cfrags attached).