Use fixed binary value for test
The use of randomly generated values for pytest parameters may cause
errors in test explorer in Visual Studio Code. Specifically,
`test_return_value_serialization` test returns the following error when
a random parameter is used:

```
Test result not found for:
./tests/unit/conditions/test_return_value.py::test_return_value_serialization
[\xbd\x01\t\xecV\x05\x04\xe2\x9b(\xb6\xcef\xea\xa4\xfa]
```

Since the rest of values used in this test are fixed, there is no reason
to use a random value for bytes parameter.