Fix Prometheus import error when running Porter
Commit a4de8a88 introduced amazing changes and solved linting errors
like F821. To achive it, this change introduced the import of
PrometheusMetricsConfig in lawful.py, which is dependant of
prometheus-client package.

Ursula imports prometheus-client, so no problem with this. But Porter
does not. So a import error is raised when Porter is run.

Also, a remaining TODO comment has been cleaned. This TODO is part of a
larger comment and was introduced in commit 424e373a and solved
afterwards.