Fix missing argument in character config loading
PR #3207 introduces a new migration script of the config JSON from v5 to
v6. This new version adds a new property "condition_providers" to the
JSON.

The configuration loader doesn't expect this new field, so an
"unexpected keyword argument'condition_provider'" is raised when trying
to run ursula `nucypher ursula run`.

This commit adds this field to the loader.