Add .ruff_cache folder to .gitignore file
Ruff linter generates a cache folder when it runs pre-commit checks.
This folder is ignored since it has its own .gitignore file inside the
folder. But for homogeneity purposes this should be added to .gitignore
file.

Also, adding to general .gitignore file makes that this folder is going
to have a different treatment by IDEs.